A js library used to change textContent.
This js library is incomplete, so do not use! Its intended use is to let you change the textContent of a webpage element in response to the user clicking a different element. To use, call the "addSet" function, with the first parameter being a string that is the ID of the element that will be translated (I will call this element "element 1"), parameter 2 being a string which will be element 1's future textContent, and the third parameter being the ID of the element that the user will click. You may call this function as many times as you like, and once you have called it as much as you want, call the function "addTranslators" to add event listeners. Warning: This is an incomplete js library with extremely limited functionality and this functionality possibly will not work as expected.